This commit is too long, but I figured I'd push it with some things untested, than keep making it larger.  Wish not to repeat this again.
- Clean the page fault handling path to actually work (fixing some
race conditions)

- add two page memory managers, the page_pool for page tables (used in
mpd), and mmap, an array of pointers to pages to be used with the
mmap_cntl syscall to grant/revoke memory pages

- add the mmap_cntl syscall with grant and revoke, but without many
practical checks (can we make a grant/revoke, are we allowed to
execute mmap_cntl, etc...)

- Add support both for the system info page and data pages.  Remove
the get_thd_id syscall instead using the system info page information.

- Enhance the loader to automatically create the capabilities from
component to component.  This was a maintenance nightmare, and made me
never want to add components.  For now, fixed, with a simple
convention for naming the server and client stubs: For function
fn_name, fn_name_inv is the stub both in the client and server.  In
the client, there is a default stub when args are all passed in
registers.

- Change the ipc path to pass 4 args in registers instead of 3.

- Fixed the reference counting for composite spds for upcall,
branding, and creating threads.

- Add some atomic ops and add support for them in the loader and the
ioctl interface, but not yet in the interrupt paths.  Include cmpxchg.

- Fill in, but don't complete the brand system call.  The code already
exists to do this in the hijack file for timer interrupts, but not as
a system call.  Adding that, but it will probably be put on hold as I
don't think it's necessary for the tests.

- Add initial support for cos_scheduler.{c|h}: lib code for
schedulers.  Need the atomic instructions to really make this work.