1) Add the notion of atomic operations allowing user-level sections of code in which upon a preemption, the kernel will see if the instruction pointer is in one of these regions, and if so, it will rollback the ip to the beginning of the region.  Allows for atomic operations without requiring atomic instructions and their overhead. Does _not_ work for multi-core/processor systems however.
2) added the block and unblock on lock operations to the
cos_thread_switch syscall allowing a scheduler to block on a
(currently single) lock.  Created cos_sched_lock_{take|release}() in
cos_scheduler.h to access this functionality.

3) Fixed the behavior of interrupt threads whereby the preempted
thread is scheduled before the interrupt thread finishes.  In this
case, an upcall is made into the scheduler component notifying it of
the termination, and allowing it to schedule another thread.