Fixed bugs related to getting the networking to work, and plugged in with brands.
- Changed cosnet so that it maintains composite state across the
lifetime of the module instead of the lifetime of the file holding
process...this makes testing much less painful as the cnet_user script
can be run in the forground and allowed to terminate.

- fixed up the pending brands handling code including hooking in the
callback to the networking module to remove a skb from the packet
queue.

- Included setting composite to the hightest priority as it seems to
hide some bugs.

- added a buglist to cronicle my pains.

(can you tell I'm tired from my spelling?)