1) fixed edge case in network handling where pending events weren't sent for packets arriving when the queue was full
2) added a deferrable server which currently just receives upcalls and
returns.

3) added functionality to the fprr scheduler to run the deferrable
server, and to exercise brands to test their performance.

4) Modified the brand_upcall syscall to work again using
brand_next_thread and co.  Possible pending bug: I don't save the
eflags of the branding thread.  If we immediately execute the brand,
and later (brand_execution_completion) return and restore the branding
thread, it will load in the eflags register from the thread structure,
which might not be valid.  Should probably just put a sanitized eflags
value in when saving branding thread's contents.

5) added the tests directory which contains data, plots, and plotting
files.