- Ioctl infrastructure for enabling and disabling system calls to ensure composite doesn't accidently call any of them.
- Add semantics for returning from a component with nothing on your
invocation stack: upcall into the base scheduler.

- Add system calls to read and reset the number of invocations between
two components.

- Add mechanism to pass the component graph to the mpd_mgr at load time.

- Add dietlibc to be compiled into each component.  This allows the
use of sprintf to setup a string, and pass it to the print component
to print out.

- Assorted flag manipulations in dietlibc to support correct string
fns and our malloc.

- Changes to fp_rr: support to pessimistic locks on a per-spd basis.
Initial support for wait-free support for dependent tasks, wakeup and
block calls with timeouts.  Support for a single timeout that will
wakeup a specific (event) thread when it expires.

- timed_event.c using the timeout capability from the scheduler,
provide a component for managing generic timeouts.

- lock.c to provide mutexes that can be timed using the timed_event
support.  These work with synchronization.h such that no contention
avoids calls to the lock component.  However, when there is
contention, 3 calls to the lock component must be made in all
(prelock, lock, and release) to avoid race conditions.

- mailbox.c provides mailboxes using the locks above.  This code is
NOT TESTED in this commit.

- Cleaned up the assembly and stubs abstracting quite a bit into
header files.  See s_stubmb.S for and example.

- added mpd_mgr.c to currently just print out and zero # of
invocations between components.

- Painfully rewrote the component Makefile so that compiling with
dietlibc worked and doesn't interfere with composite functionality.