1) added cos_spd_name_map.h to translate component strings to their id.  This is used first in cos_loader to make sure that the components and their ids match up to the list in the .h file, and (now) in fp_rr to upcall into specific components by name as opposed to by hard-coding in the id.
2) fixed up fp_rr such that the init thread is integrated into the
scheduling infrastructure.  Now booting up is cleaner.

3) added the ring_buff abstraction so as to simulate a
network-provided packet ring buffer for a network device.  A ring
buffer is attached to a (net) brand, so will contain only information
relevant to that principal.  The ring buffers are constrained in that
the buffers they point to much not overlap pages, and can have at most
512 (pagesize/entrysize(=8)) packets pointed to in the ring.  Brand
support in the networking code locates the appropriate user-level
buffer on brand.  This eliminates the need for the ugly code to
try_brand... called a few times in awkward places.  In the networking
code (cosnet.c), the real skb_buff gets deallocated immediately as
opposed to being queued in linux.  This is beneficial as it moves all
that policy to components.