The main purpose of this patch is to provide buffered communication with the network for threads.  Threads can read and write from packet queues that are added to by the lwip layer.  Currently working (mostly) is UDP receive.  Still need to test UDP send, and get TCP working.  This was one hell of a patch to debug:
1) changes to cos_alloc to add lock-free manipulations to the free
lists (still need to make lock-free the heap ptr.

2) changes to mem_man.c to fix an idiotic bug that made hardware
aliaes for subsequent calls to map a page.  Ouch.  Support for aliases
does not exist in the API as of now.

3) fp_rr to clarify with some of the asserts that even threads that
are not THD_READY, can make calls we would typically associate with
ready threads.  This is due to upcalls waking asynchronously.  Also
fixed up the sched_block call to take into account additional
edge-cases.