1) Add reasonable run-time monitoring for each thread in fprr.  A good debug aid.
2) Fix a bug that was eminent: when an upcall finishes, it will
automatically switch to the next thread in its preemption chain
without scheduler involvement.  This is undesirable when that upcall
thread wakes up a middle priority thread and the preempted thread is
the idle thread (low prio).  Now when the upcall completes, it will
automatically switch to the idle thread.  Oops.  Added a flag for
cos_sched_cntl to remove the current thread from the preemption chain.
Hugely improved latency.