NOTE: This commit is broken.  only with the next commit will it work. (forgot to add a file)
This adds support for an event component, and makes the appropriate
changes to the networking component to use that event server.

Add a file descriptor api that defines the blocking semantics. (next
commit)

Add a connection manager that reads and writes, and tests all
appropriate functionality.

UDP and TCP connect have NOT been changed to use the event server.