Add more counters in the kernel to monitor upcall behavior. Additionally, add support for statistical counters both in the kernel and in the cos_net component.  Add support in the kernel for measuring the latencies to go from brand-> upcall, pending brand-> upcall, and brand-> upcall done.
Add a system call to print out the statistics and counters.

Add an additional priority to the event component so that some events
will never be triggered when others are available.  Currently only 2
priorities are provided, but the code is fairly general.

Remove the resume_return system call from inv.c.  Hasn't been used in
a long time, and is probably wrong by now.

Changes the scheduling semantics s.t. when an event is triggered from
the kernel into the shared scheduling region, it clears a bit.  If
this bit (next thread) isn't set on switch_thread, then return and
make sure that the scheduler notices any events.

Remove the "urgency" argument to switch_thread as it is ignored.

Change cos_net to use the cos_map utility.

Fix the idle processing of scheduler events.

Change components that used a bootup thread then called sched_block
indefinately.  Instead, they let the thead die and it is collected by
the scheduler.