1) add "open" for content, 2) fix routing for content mgr, 3) make events their own namespace
Make opening a content request manditory and add support for this in
https, content_mgr, and static_content.

Fix the routing algorithm in the content_mgr component.

Make events be identified by their own namespace (managed in evt.o),
instead of by an external identifier such as file descriptor.  This is
necessary to allow the same evt component to be shared across
independent components.  The evt component should perform some
security checks to make sure that triggering can be done from the
calling component, but this is not currently done.