Main point: change from using upcalls to synchronous event threads.
1) Previously, we had upcall threads associated with brands.  Now we
have synchronous threads that can "wait" on brands for a brand
activation.  This allows invocations between components for
upcall/event threads to progress exactly as normal invocations.  This
has the benefit of falling into the current structure easily (for
free!), but the detriment of possibly keeping a stale MPD around for
longer.  Such is life.

2) Add an event monitoring infrastructure into the kernel that records
scheduling/upcall events into a ring buffer and the ring buffer is
output when the module is removed.  This makes debugging problems
associated with scheduling and events much easier.

3) remove comp[123].c

Thus concludes the life of this (successful) branch!