Add the split method into the async_inv component; integrate with cgi
Add the split method which is analogous to accept with sockets in that
it is executed on an already open descriptor, and yields a new
descriptor (thus split).  In this context (async_inv), the cgi script
opens a single main descriptor that designates its service name, and
it splits off requests, each of which is answered individually.  This
makes it clear when a response is fully formulated as the descriptor
is closed by the cgi script.