1) Reorg the Makefiles including moving declarations into Makefile.inc
2) Initial idle implementation including syscall.  Changes to inv and
hijack.  Not currently completely functional.  Will hard crash easily
if pushed.

3) Fixed bug interfacing with the compiler in the initial mapping of
the shared data page.

4) Addeded an event notification bit into the scheduler notification
page to signify that an event (upcall activation, or addition to the
evtlist) has occured.  Avoid races -> want sched -> about to
switch_thread -> event -> continue switching, missing event.  Not yet
integrated with child/parent interaction.

5) Consolidate all cos_switch_thread calls in fp_rr into a single
function that deals with all error cases in a uniform manner and kills
a lot of code.

6) enable 32 rather than 16 syscalls.

7) Fix bugs in the loader dealing with changing bfd versions.  Remove
assumptions on presence of sections.

8) add the sched_ops structure for abstracting away scheduler
functionality, and implement a fp_ops, for the fixed priority runtime.