1) expanded stack tracing so that the current thread can trace it's steps, and fixed a condition for finding the stack pointer in a thread in a separate protection domain in the process of making an invocation.
2) add support for idle in linux.  The idle system call will call
out to Linux and put the composite thread on a wait queue.  The
interactions between this and the interrupt/brand subsystem is
complicated.  Idle is broken into three states: awake (in Linux
runqueue), asleep (on the hijack waitq), and waking.  We know to
remove the thread from the runqueue if it is asleep, but not if
it is waking.

3) fixed a bug in the interrupt/branding code where registers
were being manipulated at the top of the thread's stack when they
should have been.  This is the long-standing eip=0x10 bug.
Concurrently fixed a bug resulting from nested interrupts and the
idle state (which is what made idle so complicated).

4) add more monitoring infrastructure.

5) print out minimal component information on
execution (including base address, and component id).

6) Abstract out the implementation of most of the functionality
for specific schedulers from the basic scheduling toolkit base.

7) consolidate all the difficult switch thread bits into the
sched_switch_thread function.  This includes moving the loop (on
error, or switch-thread-self) and all dependency processing into
a single location.  This is the best change I've made to the
scheduler so far.  This touches a lot of code in fp_rr.c.