Cleaned up hijack.c to remove dead/old code, and fixed page_fault interposition
Removed much of the aed (asymmetric execution environments) code from hijack.c.

Cleaned up the code in kern_entry.S to be easily extendable to
arbitrary hardware entry points.

Fixed (finally) the bug with page fault handling that precluded us
interposing.  Module code being vmalloced and on-demand faulted in,
could not define the page fault handler (doing so resulted in double
faults).  Prevent this by defining a low-mem trampoline interposer
that will go to the linux handler when the fault is in the
kernel...otherwise enable our usual interposition.  Fricking pain in
the ass.  2 weeks to determine what was causing the problem.  8 hours
to code up the solution -- thus is the way of systems.