Add ability to run most of the system in a child scheduler.
1) Separate mpd mgr into the actual manager, and the capablity graph
(cgraph) storage.  The first is can now be booted and stkmanaged, the
second cannot, but can be added to by the booter, and can be queried
by other components.

2) The boot/loader component no longer creates all threads for its
components.  Components under a child scheduler will be created by
that scheduler directly.  The scheduler must be initialized _after_
the booter/loader runs.

3) Added lws_hier.sh that boots the webserver in a child scheduler.

4) Fixed a bug in the scheduler to ensure that child "timer" threads
are not added into runqueues.

NOTE:  I forgot to add the hw_ints.[ch] last time.  Added this time.