Stack misses take into account thread dependencies now.
1) ensure that stacks hold the thread id of a thread using the stack

2) in the stack manager, make sure that we use the sched_block with
dependency support to provide priority inheritance of current thread
causing a stack miss to the the thread holding the stack.

3) take into account thread suspensions wrt the dependencies created.

TODO: 1) if we attempt to make a dependency on a thread that is
waiting for a stack in a component with self-suspended threads, then
we should increase the self-suspension count of the component that
actually has the threads with the self-suspensions...not the one we
are stack missing in.  2) add support in the stack manager/policy to
recognize self-suspension stack usage, and allocate stacks to
components with SRT threads where the components have self-suspensions
with priority to those that do not have self-suspensions.