Fixed assert error in cos_sched_base which is caused by executing pre_block more than one time. Ensure each component at least owns one stack in stack manager.
1) when we try to depend on other thread, we would execute pre_block
which is the same as usual blocking. But we are not sure if we should
be modifing wake_cnt here. Now the problem is if we depend on other
threads more than once, the wake_cnt would be less than 0. We fixed it
by only decreasing wake_cnt if it is greater than 0.

2) Record a number of "empty components" in stack manager. And ensure
the remaining stacks number is at least equal to the number of empty
components. This can ensure each component has at least one stack
quota. Also modified stack policy to ensure enough stacks for time
event component(for temporal stack test).