Implemented self-suspension components detection. Changed both Stack Manager and Stack Policy to ensure allocating enough stacks to self-suspension components.
1) Add a self-suspension counter to each component in Stack
Manager. Increase the counter when we can't find any thread to depend
on, which means all threads are self-suspended and this component is
self-suspension related (Not necessary to be where exactly
self-suspension happened).

2) Allow self-suspension over-quota allocation privilege. Will ignore
the quota limit of self-suspension components when granting stacks to
them.

3) Add a new interface to stack manager(detect_self_suspension). It
returns the self-suspension counter of the component, where a positive
value means there is self-suspension and 0 means not. Stack policy
uses this interface to gather information about self-suspension. And
policy ensures enough quota for those components.

4) Fixed a bug: pre_wakeup should be called if a thread is waken up
due to its dependency thread self-suspended.