Main memory mapper now is synchronized.
The main memory mapping component (using the improved mapping
data-structure) is now synchronized.  This was difficult as it had to
become the root scheduler so that it could use the synchronization
facilities of a scheduler, while still enabling the real root
scheduler to still maitain the ability to e.g. create threads.

The main memory mapper is now in a great state with very few FIXMEs to
be done aside from fixing the inevitable bugs.