Rewrite of the cbuf client-side code.  Removed the slabs, and simplified the logic.  Now all allocations are the size of a page.  In the future, we will include multi-page requests, but never sub-page. That can be provided by user-level libraries if required.  Please run unit_cbuf.sh to test.
This provides a good study in removing complexity.

TODO includes adding microbenchmarks to the cbuf unit tests.  Given
the reduction in complexity, the use of cvect where possible, and the
removal of many data-structure walks in the cbuf_vect, I'd expect this
to be roughly 4 times faster than the previous code.