Scheduling of the creation of threads for components on boot-up.  If you want an initial thread, then simply depend on the scheduler. WARNING: breaks backwards compatibility in that thread priorities are now ignored.  If you want to create a thread with specific priorities, call sched_create_thread explicitly.
Added a new data-area that is used to pass the schedule to the booter,
and it creates threads in this order.  We wanted to maintain some
backwards compatibility: Components that wish to use the boot up
thread to continue computation can do so.  The thread within the
booter is at a lower priority, so if the initialization thread in the
new components blocks, only then will the next thread be created.
This does imply that if you block during initialization, then you
could recieve invocations after that point.  This seems like a
practically OK trade-off for now (i.e. most initialization should be
done without blocking).