Found TLB coherency issue, and hacked a fix that needs to be fixed later.
In the booter, when we go beyond the initial 4MB region, the next page
that is touched (that has already been mapped) will fault with a "page
not present" fault.  The page tables are setup fine, both at the pgd
and the pte level.  Flushing the TLB solves the issue (thus I added a
COS_MMAP_TLBFLUSH flag).  Currently, the tlb is flushed after the
pages are mapped in, but before they are accessed, and it works.  The
only way this really makes sense is if invocations aren't going
through the kernel (i.e. which should flush the TLB).  However, even
in that case, there should be no TLB mapping for the missing page that
is not active, thus the TLB should not include a "not present" page.
To make it more confusing, the system works for any allocations made
within the same pgd (4MB region) -- thus the "in same protection
domain" theory is wrong -- and it only fails when we go to another pgd
entry.

Who knows?