Added the evt_split function to the event interface, a new edge-triggered event manager that is simpler, and provides a better interface for waiting on multiple events, and changed the runscripts to use it.
Specifically: When a new event is created, you must specify (using
evt_split) which other event is this new event's parent, and if this
event is a "group" or a normal event.  Group events can have new
events split off of them, and we can wait on the group.  This will
yield wait returning the event id of the child event that triggered
the event.  A tree of these groups can be defined.

Warning: The asymtotic complexity of these calls is proportionate to
the _depth_ of the group hierarchy.