Implemented a new intelligent waking up scheme in the tmem manager.
The new waking up procedure is: 1) low thread release the lock, then
wake up the highest thread, which is the last one on the block
list. 2) the highest thread wake up the rest ones on the block list,
without releasing the lock.

A epoch counter is added to determine whether a thread should be wake
up by the current thread.

Some other fix: fixed the priority of the timer thread in timed event
component. fixed the priority issue in the synth hier component.