Step 1 toward a low-level booter (llboot).
The goal of this string of commits is to provide a low-level boot
component that will load _all_ other components (including the normal
boot component).  The has three purposes:

1. Since only a single component is required to boot an arbitrarily
   large system (given enough memory for that single component), it
   becomes easier to boot from the bare-metal as during system boot
   up, we only need to create a single component without any
   capabilities.  A later step will include serializing this component
   image to disk so that it can be loaded as a ramdisk via the
   multi-boot standard (i.e. as a module).

2. Fault tolerance: If we want to be able to recover from faults in
   the lowest-level components such as the memory manager and the
   scheduler, then we have to have some component that can trap the
   errors, and reboot the offending component.  This is the low-level
   booter.  Support for e.g. fault handling will be included in a
   later commit.

3. To provide a low-level execution environment for the most
   timing-critical of components.  When abstractions such as the
   scheduling layer, the memory manager, file systems, stack
   management, event management, etc... entail too much overhead, and
   an application wants full access to the hardware, it can be booted
   by the low-level boot component.

Care was taken to keep the llboot component relatively small.
Discounting data-structures, and other generic code, the component is
less than 500 lines of code, and is very specialized around its
purpose.  In spite of this, it does share the majority of its code
with the normal boot component.

Currently the system does _not use the llboot_ component.  This commit
maintains backwards compatibility with existing runscripts, while
changing the way that scheduler intialization works (so that it can be
done for the root scheduler via upcall), and adds the necessary
support in the llboot component for future functionality.  To do the
latter, I had to separate out some of the normal and current
functionality for the boot component into boot_dep.h; a change that
is, again, backwards compatible.