Added the get_cpuid_fast() function, which is only used in the invocation path to get the current cpuid while optimizing for uniprocessor performance.
The reason that we have a seperate _fast function is because:
(Complicated...)

1) We want to optimize the uniprocessor case, which means we should
have the get_cpuid function returns a constant 0 when cpuid == 1.

2) The get_cpuid function is also used by the hijack module and

3) We need to do hijacking on all cores, even Composite only uses a
single core. (Want to change this!)

As a consequence, when NUM_CPU is 1, the get_cpuid function cannot
both be optimized for the uniprocessor(return 0) and reflect actual
CPU id in the hijack code for cores other than 0. Thus we have the
get_cpuid function to return actual value always, and the _fast
function to optimize for performance. Currently it's only used in the
invocation path.