First phase of loading improvements:  partial cos_loader->generic section loading + non-partitioned mem{cpy|set} to initialize components.
1) First phases of adding a generic mechanism for adding new sections
to the loading process.  Will be used in phase 3 (thereabouts) to add
initonce sections, and in phase 4 to add a kernel-mapped section.
Later, we can add the .cons and .dtor sections, etc...

2) In the composite booter, we used to do memcpy and memset page at a
time.  This slowed down the processing as it messed with the
processor's prefetcher.  Much better to simply memcpy or memset the
entire section we are trying to initialize.  This is done in this
patch.  Note that one wouldn't typically care to optimize this, but
when we want to do fast component micro-rebooting, it matters.

3) I moved the cos_component.c library file from a library-loaded
file, to a manditory-included file.  We assume that all components
have a cos_comp_info page.  TODO: remove the scheduler notification
page from this, and place it in scheduler-specific manditory-include
library files.  This will save a page per component (of kernel-shared
memory).