Remove the scheduler notifications page from non-schedulers.
Previously the `cos_sched_notifications` page was a manditory part of
every component.  This blew up the BSS by a page.  Not the end of the
world, and we've avoided caring about it up till now.  However, this
will add the cost of zeroing out a page for recovery of
non-schedulers.  We'd like to avoid this if possible.  This commit
removes it, and instead adds a `cos_sched_ds.h` file that includes the
data-structure itself.  Almost no component should have to include
this manually as if you're writing a scheduler, you should inherit the
scheduler interface which will include it automatically.