Added new object sections including ctor, dtor, crecov, and cinfo.  Also included support for using these sections (i.e. placing functions pointers in them).
See cos_component.h for CCTOR, CDTOR, and CRECOV to set a function as
belonging to one of these classes.  If it is constructor, then it will
be called on bootup before cos_init.  If it is a destructor, it is
currently not called.  Jiguo will need to add support so that when a
recovery upcall is called, it will invoke the recovery handlers
(i.e. this is used as a notification of needed recovery.

If you define your own cos_init, then you should call
constructors_execute yourself.