Unit Tested captbl done.
Has multi-sized capabilities, and checking to make sure that the
correct size is used, and that allocated vs. freed is accurately
tracked.  Added code to the capability translation to return NULL or
the pointer depending on if it is allocated or not.  This uses
branch-free code based on bit operations, assuming twos complement
(i.e. 0-1 = MAX_UINT).