cbufp: break dependency with cbuf
Break the dependency of using a cbuf to return the garbage-collection list
from cbufp_collect() by mapping a shared page between the cbufp manager
and each client component. The list is stored in a ring buffer that is
accessed via a ck_ring structure located at the start of the shared
page.