Added the u64_t deact_tsc field to the capability header: it's the time stamp for the deactivation time of the capability entry. Since we need it for all types of capabilities (even for core-local caps like threads), it's integrated in the header.
The 64-bit tsc changes the sizes of most types of capabilities. Fixed
many places that assume the size of capability entries.

Todo: a cap id allocation mechanism in the llboot for the low-level
components loaded by llboot.

Possible issue: we have a deact_tsc in each cap entry. However, when
we re-use the cacheline for a different type of capability, we may
need to check all TSCs in the cacheline.