cos_loader: enable additional boot layers
Call make_spd_boot on subsequent boot layers past 1. The
booters need to be invoked in reverse order in case their
component memory changes. Also create the boot schedule
for each layer, and assume that dependencies are satisfied
if another booter should load it. Note: the logic for
checking dependencies in spd_already_loaded() should be
made more robust by checking if the service is loaded by
an earlier layer in the boot sequence.