cbboot: add a new component booter intended to nest below boot
The cbboot currently uses the same primary mechanisms as boot,
but its goal eventually is to use cbufs as the backing memory
for components that it boots. The unit_cbboot_cbuf.sh script
is identical to unit_cbuf.sh except for using the cbboot to load
the test component objects.