cbuf: add a replacement SS_ipc_client_fault
Make the default SS_ipc_client_fault weak, and add a strong version
in the cbuf interface. Add a new test pftest_cbuf that is like pftest
but pulls in cbuf. Cause the cbuf version of SS_ipc_client_fault to
return a different expected value, and check for that in pftest_cbuf.