Progress in booting up the low-level booter.  Gets to "done creating system".
Bug fixes:

Fixed a bug to get debugging flags working.  CK uses a __builtin* for
bit population counting, and its implementation in libgcc uses
_GLOBAL_OFFSET_TABLE_ (i.e. it is PIC).  The "fix" is to comment out
the usage of the builtin function, and cause a fault instead.

Fixed a bug due to incorrect heap meta-data in the llbooter object
resulting in faults upon boot.  A simple update of a heap pointer
fixed that.