Modify cos_rcv to only return a 1 or a 0
While this goes hand in hand with the path to fix concurency on qemu,
this is a stand alone patch as it should be considered a "hot fix" of
sorts. As a stand alone commit it is easier to revert this commit if
neccessary.

The code was modifed such that the cos_sync library could act on all
pending events. There was a bug where the number of pending events would
never be 0. If a revert is needed, a more thorough debugging of the api
may be necessary.