Removed old logic for tcap_transfer
We do not use the old version of transfer anymore.  It required that the sender have a uniformly higher quality than the receiver.  This seems impossible to maintain in a scheduler that coordinates at all with other schedulers.  So all places in the code that used transfer have been moved over to delegate.

- Move the last use of tcap_transfer over to tcap_delegate (in tcap_merge)
- Remove the code for tcap_transfer

TCap code down to < 250 LoC.  Great!