Initial commit with budget tracking and timer vectoring.
Implements the use of one-shot timers for all time-keeping in the
system.  This means that when switching between threads, we can set
the next timeout, and if a value isn't provided, or if an asnd
activates a thread, then we use the tcap budget to set the timer.

This code *does not* fully support budget tracking and interrupt
vectoring as will need to be implemented in the final system.

It does:

- micro_booter `test_timer` uses tcap budgets and thread switch timeouts
- add a chal interface for setting the oneshot timer
- add a time API for the micro-booter that enables the specification
  of timeouts on an absolute timeline, with conversions to and fro cycles
- budget accounting/consumption on tcaps

It does not:

- budgets < INF (lack of testing)
- do accurate accounting...it seems that the timeout math is
  triggering timeouts that are 33% too short, or so
- proper switching logic if we're changing tcaps
- efficiently set the timer: it only correctly sets the oneshot by
  calling the `set` function two time successively...and I have no
  idea why.
- include logic for dispatching between tcaps with budget, instead it
  focuses on dispatching only to the current tcap's parent (see INF
  budget, above)