Change the src Makefile to use PLATFORM_ID instead of PLATFORM
The src Makefile creates a temporary file that stores the name of
the platform the build is running on. (It does this when you run
`make config`.) This file was historically called PLATFORM, which
worked fine. However, there is a folder in the src directory
called `platform`. On case insensitive file systems this causes an
obvious problem. To fix this, I modified the Makefile to use
PLATFORM_ID instead.