Remove duplicated call to OS_API_Init, preventing a fault
When running unit tests OS_API_Init was called twice, leading
mem_info to be propogated twice. This lead to a fault when
allocating after the second call, since the memory being allocated
was already in use.

This fix deduplicates calls to OS_API_Init (a preferable solution
to changing the unit tests.)