Fix hardcoded stack size and alignment
Thread stack size is supposed to be determined by a constant in
consts.h. A lot of the code in the codebase relies on this,
including the code for fetching the a thread's id. However, the
assembly code that was really allocating the stack was using a
hardcoded value.

This mismatch led to invalid thread id's being returned, and
broke increasing stack sizes to avoid stack overflows.

To fix this, I made the assembly dependent on #defines, and made
sure those defines were available for those files. I also fixed
how those defines were structured in the first place.