Fix no-free-ID error on uninitialized queues
The `queue` struct had a field `free` that denoted whether or not the
queue at the ID was in used. Because there is no function to initialize
the queue array, all queues started with a `free` value of `FALSE`.
The `free` field has been replaced with a `used` field, and all boolean
checks have appropriately been inverted.