Store symbol names in `cobj`
The original format of `cobj` only stores the address of a symbol along
with an index. The OSAL needs to be able to look up symbols by name, and
@gparmer and I have decided that the best way to parse the objects would
be to use Composite's existing `cobj` parsing library rather than
writing an ELF parser.

This commit adds a new field to the `cobj_symb`, `name`. In the future,
this implementation will avoid wasting memory by putting symbol names at
the end of the cobj.