Update filesystem including good filedescripters, permissions, and file commands
The filesystem passes all of the cFE unit tests, it provides all of the
functionality that is required by the cFE_booter and that will be
needed by the linking and loading.  It is much closer to POSIX in the
way that is uses FDs.  It is still not threadsafe or modularized enough
to create a seperate library.